login authentication

+++ app/Http/Controllers/TasksController.php
@@ -4,9 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Http\Requests;
+use App\Http\Controllers\Controller;
 
 use App\Task;
+
 
 class TasksController extends Controller
 {
@@ -17,11 +18,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all();
+         $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -48,14 +58,13 @@ class TasksController extends Controller
     {
          $this->validate($request, [
             'content' => 'required|max:191',
-              'status' => 'required|max:10',
-            
+             'status' => 'required|max:10',
         ]);
-        
-         $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -68,11 +77,15 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-         $task = Task::find($id);
+        
+        $task = Task::find($id);
+
+        if (\Auth::user()->id === $task->user_id){
 
         return view('tasks.show', [
             'task' => $task,
-        ]);
+        ]);}
+        return redirect('/');
     }
 
     /**
@@ -84,10 +97,12 @@ class TasksController extends Controller
     public function edit($id)
     {
          $task = Task::find($id);
-
-        return view('tasks.edit', [
+          if (\Auth::user()->id === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -119,14 +134,13 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-         $task = Task::find($id);
-        $task->delete();
+    { $task = Task::find($id);
+
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
-    public function sample() {
-  $status = config('status');
-  return view('sample')->with(['status' => $status]);
-}
+    
 }
